+++ src/pages/mypage/index.tsx
@@ -1,4 +1,4 @@
-import MypageHeader from './mypage-header'
+import MypageHeader from './components/mypage-header'
 
 export default function MyPage() {
   return <MypageHeader />
